+++ apps/control/src/utils/scheduling/MaterialTableUtils.ts
@@ -1,5 +1,6 @@
 // Columns 정의
 import { ColumnDataType } from '@postcoil/ui/config/TableConfig';
+// import { Tag } from 'antd';
 
 import { MaterialDataType } from '@/config/scheduling/ContentConfig';
 
@@ -8,13 +9,9 @@ export const mockcolumns: ColumnDataType<MaterialDataType>[] = [
     title: '재료번호',
     dataIndex: 'materialNumber',
     key: 'materialNumber',
-    // fixed: true,
+    fixed: true,
   },
-  // { title: '재료진행상태', dataIndex: 'materialStatus', key: 'materialStatus' },
-  // { title: '조업코드', dataIndex: 'operationCode', key: 'operationCode' },
   { title: '공정코드', dataIndex: 'processCode', key: 'processCode' },
-  // { title: '재료유형', dataIndex: 'materialType', key: 'materialType' },
-  // { title: '재료진도', dataIndex: 'materialProgress', key: 'materialProgress' },
   {
     title: '재료단위',
     dataIndex: 'materialGroup',
@@ -35,18 +32,6 @@ export const mockcolumns: ColumnDataType<MaterialDataType>[] = [
         record.materialGroup.indexOf(value) === 0,
     },
   },
-  // {
-  //   title: '외경',
-  //   dataIndex: 'outerDiameter',
-  //   key: 'outerDiameter',
-  //   sortable: true,
-  // },
-  // {
-  //   title: '내경',
-  //   dataIndex: 'innerDiameter',
-  //   key: 'innerDiameter',
-  //   sortable: true,
-  // },
   { title: '폭', dataIndex: 'width', key: 'width', sortable: true },
   {
     title: '길이',
@@ -66,16 +51,6 @@ export const mockcolumns: ColumnDataType<MaterialDataType>[] = [
     key: 'unitWeight',
     sortable: true,
   },
-  // {
-  //   title: '통과공정',
-  //   dataIndex: 'passedProcess',
-  //   key: 'passedProcess',
-  // },
-  // {
-  //   title: '전공장공정',
-  //   dataIndex: 'previousProcess',
-  //   key: 'previousProcess',
-  // },
   {
     title: '차공정',
     dataIndex: 'nextProcess',
@@ -104,11 +79,14 @@ export const mockcolumns: ColumnDataType<MaterialDataType>[] = [
     title: '목표폭',
     dataIndex: 'targetWidth',
     key: 'targetWidth',
+    sortable: true,
+    defaultSortOrder: 'descend',
   },
   {
     title: '목표두께',
     dataIndex: 'targetThickness',
     key: 'targetThickness',
+    sortable: true,
   },
 ];
 
@@ -125,7 +103,7 @@ export const mockdata: MaterialDataType[] = [
     materialGroup: 'A',
     outerDiameter: 35.9824301584831,
     innerDiameter: 22,
-    width: 790,
+    width: 775,
     length: 2547,
     thickness: 0.621409153683,
     unitWeight: 68.5737101624795,
@@ -149,7 +127,7 @@ export const mockdata: MaterialDataType[] = [
     materialGroup: 'A',
     outerDiameter: 52.9602313116795,
     innerDiameter: 25,
-    width: 788,
+    width: 772,
     length: 6848,
     thickness: 0.504473434969,
     unitWeight: 434.688234352393,
@@ -157,7 +135,7 @@ export const mockdata: MaterialDataType[] = [
     previousProcess: '1PCM',
     nextProcess: '포장',
     changed: true,
-    targetWidth: 788,
+    targetWidth: 772,
     targetThickness: 0.49,
   },
   {
@@ -192,14 +170,14 @@ export const mockdata: MaterialDataType[] = [
     materialGroup: 'A',
     outerDiameter: 54.0121547172835,
     innerDiameter: 28,
-    width: 757,
+    width: 769,
     length: 6702,
     thickness: 0.337581205233,
     unitWeight: 666.443669362022,
     passedProcess: '1CAL',
     previousProcess: '1PCM',
     nextProcess: '포장',
-    targetWidth: 750,
+    targetWidth: 765,
     targetThickness: 0.3,
   },
   {
@@ -213,14 +191,14 @@ export const mockdata: MaterialDataType[] = [
     materialGroup: 'A',
     outerDiameter: 60.8096102448159,
     innerDiameter: 28,
-    width: 759,
+    width: 766,
     length: 9154,
     thickness: 0.50512372108997,
     unitWeight: 504.379666723385,
     passedProcess: '1CAL',
     previousProcess: '1PCM',
     nextProcess: '1EGL',
-    targetWidth: 750,
+    targetWidth: 765,
     targetThickness: 0.39,
   },
   {
@@ -255,14 +233,14 @@ export const mockdata: MaterialDataType[] = [
     materialGroup: 'A',
     outerDiameter: 42.1349622390059,
     innerDiameter: 21,
-    width: 800,
+    width: 774,
     length: 4192,
     thickness: 0.51703509387288,
-    unitWeight: 122.689475058989,
+    unitWeight: 122.689476558989,
     passedProcess: '1CAL',
     previousProcess: '1PCM',
     nextProcess: '포장',
-    targetWidth: 788,
+    targetWidth: 772,
     targetThickness: 0.47,
   },
   {
@@ -276,14 +254,14 @@ export const mockdata: MaterialDataType[] = [
     materialGroup: 'A',
     outerDiameter: 53.3285872359957,
     innerDiameter: 25,
-    width: 790,
+    width: 775,
     length: 6971,
     thickness: 0.714454085291,
     unitWeight: 417.961959629394,
     passedProcess: '1CAL',
     previousProcess: '1PCM',
     nextProcess: '포장',
-    targetWidth: 788,
+    targetWidth: 772,
     targetThickness: 0.58,
   },
   {
@@ -304,7 +282,496 @@ export const mockdata: MaterialDataType[] = [
     passedProcess: '1PCM',
     previousProcess: '1CAL',
     nextProcess: '1EGL',
-    targetWidth: 750,
+    targetWidth: 765,
+    targetThickness: 0.67,
+  },
+  {
+    key: '10',
+    materialNumber: 'Piy0010',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'B',
+    outerDiameter: 40.344223871778,
+    innerDiameter: 30,
+    width: 737,
+    length: 2286,
+    thickness: 0.86932033936218,
+    unitWeight: 33.9806455809316,
+    passedProcess: '1EGL',
+    previousProcess: '1CAL',
+    nextProcess: '1CGL',
+    targetWidth: 735,
+    targetThickness: 0.7,
+  },
+];
+
+export const schMockcolumns: ColumnDataType<MaterialDataType>[] = [
+  {
+    title: '재료번호',
+    dataIndex: 'materialNumber',
+    key: 'materialNumber',
+  },
+  // { title: '스케줄명', dataIndex: 'scheduleName', key: 'scheduleName' },
+  { title: '공정코드', dataIndex: 'processCode', key: 'processCode' },
+  {
+    title: '재료단위',
+    dataIndex: 'materialGroup',
+    key: 'materialGroup',
+    sortable: true,
+  },
+  { title: '폭', dataIndex: 'width', key: 'width', sortable: true },
+  {
+    title: '길이',
+    dataIndex: 'length',
+    key: 'length',
+    sortable: true,
+  },
+  {
+    title: '두께',
+    dataIndex: 'thickness',
+    key: 'thickness',
+    sortable: true,
+  },
+  {
+    title: '차공정',
+    dataIndex: 'nextProcess',
+    key: 'nextProcess',
+    sortable: true,
+    otherProps: {
+      filters: [
+        {
+          text: '1CGL',
+          value: '1CGL',
+        },
+        {
+          text: '1EGL',
+          value: '1EGL',
+        },
+        {
+          text: '포장',
+          value: '포장',
+        },
+      ],
+      onFilter: (value: string, record: { nextProcess: string[] }) =>
+        record.nextProcess.indexOf(value) === 0,
+    },
+  },
+  {
+    title: '목표폭',
+    dataIndex: 'targetWidth',
+    key: 'targetWidth',
+    sortable: true,
+    // defaultSortOrder: 'descend',
+  },
+  {
+    title: '목표두께',
+    dataIndex: 'targetThickness',
+    key: 'targetThickness',
+    sortable: true,
+  },
+  {
+    title: '작업소요시간',
+    dataIndex: 'worktime',
+    key: 'worktime',
+    sortable: true,
+  },
+  {
+    title: 'Reject여부',
+    dataIndex: 'reject',
+    key: 'reject',
+    // TODO: tag로 reject 여부 디자인!
+    // render: (_, record) => {
+    //   let color = 'green';
+    //   let tag = '';
+    //   console.log(record);
+    //   if (record.reject === 'T') {
+    //     color = 'volcano';
+    //     tag = 'reject';
+    //   }
+    //   return (
+    //     <Tag color={color} key={color}>
+    //       {tag.toUpperCase()}
+    //     </Tag>
+    //   );
+    // },
+  },
+];
+
+// Table Data 정의
+export const schMockdataA: MaterialDataType[] = [
+  {
+    key: '1',
+    materialNumber: 'C3Lp0008',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 53.3285872359957,
+    innerDiameter: 25,
+    width: 775,
+    length: 6971,
+    thickness: 0.714454085291,
+    unitWeight: 417.961959629394,
+    passedProcess: '1CAL',
+    previousProcess: '1PCM',
+    nextProcess: '포장',
+    targetWidth: 772,
+    targetThickness: 0.58,
+    worktime: 35,
+  },
+  {
+    key: '2',
+    materialNumber: 'COz0007',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 42.1349622390059,
+    innerDiameter: 21,
+    width: 774,
+    length: 4192,
+    thickness: 0.51703509387288,
+    unitWeight: 122.689476558989,
+    passedProcess: '1CAL',
+    previousProcess: '1PCM',
+    nextProcess: '포장',
+    targetWidth: 772,
+    targetThickness: 0.47,
+    changed: true,
+    worktime: 48,
+    reject: 'T',
+  },
+  {
+    key: '3',
+    materialNumber: 'C3DL0002',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 52.9602313116795,
+    innerDiameter: 25,
+    width: 775,
+    length: 6848,
+    thickness: 0.504473434969,
+    unitWeight: 434.688234352393,
+    passedProcess: '1CAL',
+    previousProcess: '1PCM',
+    nextProcess: '포장',
+    targetWidth: 772,
+    targetThickness: 0.49,
+    worktime: 29,
+  },
+
+  {
+    key: '4',
+    materialNumber: 'C6EK0001',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 35.9824301584831,
+    innerDiameter: 22,
+    width: 775,
+    length: 2547,
+    thickness: 0.621409153683,
+    unitWeight: 68.5737101624795,
+    passedProcess: '1CAL',
+    previousProcess: '1PCM',
+    nextProcess: '포장',
+    storageLocation: '7-2-20-2-1',
+    yardDivision: '1A1B',
+    orderNumber: '20240730-F6BG',
+    targetWidth: 770,
+    targetThickness: 0.48,
+    worktime: 34,
+  },
+  {
+    key: '5',
+    materialNumber: 'C3i00003',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 49.4609835245515,
+    innerDiameter: 24,
+    width: 772,
+    length: 5876,
+    thickness: 0.620238096071,
+    unitWeight: 645.773619063763,
+    passedProcess: '1CAL',
+    previousProcess: '',
+    nextProcess: '포장',
+    targetWidth: 770,
+    targetThickness: 0.57,
+    worktime: 38,
+  },
+
+  {
+    key: '6',
+    materialNumber: 'Cbq0005',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 60.8096102448159,
+    innerDiameter: 28,
+    width: 766,
+    length: 9154,
+    thickness: 0.50512372108997,
+    unitWeight: 504.379666723385,
+    passedProcess: '1CAL',
+    previousProcess: '1PCM',
+    nextProcess: '1EGL',
+    targetWidth: 765,
+    targetThickness: 0.39,
+    worktime: 35,
+  },
+  {
+    key: '7',
+    materialNumber: 'C3Nj0004',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 54.0121547172835,
+    innerDiameter: 28,
+    width: 769,
+    length: 6702,
+    thickness: 0.337581205233,
+    unitWeight: 666.443669362022,
+    passedProcess: '1CAL',
+    previousProcess: '1PCM',
+    nextProcess: '포장',
+    targetWidth: 765,
+    targetThickness: 0.3,
+    worktime: 30,
+  },
+];
+
+// Table Data 정의
+export const schMockdataA2: MaterialDataType[] = [
+  {
+    key: '1',
+    materialNumber: 'C3Lp0008',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 53.3285872359957,
+    innerDiameter: 25,
+    width: 775,
+    length: 6971,
+    thickness: 0.714454085291,
+    unitWeight: 417.961959629394,
+    passedProcess: '1CAL',
+    previousProcess: '1PCM',
+    nextProcess: '포장',
+    targetWidth: 772,
+    targetThickness: 0.58,
+    worktime: 35,
+  },
+  {
+    key: '2',
+    materialNumber: 'COz0007',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 42.1349622390059,
+    innerDiameter: 21,
+    width: 774,
+    length: 4192,
+    thickness: 0.51703509387288,
+    unitWeight: 122.689476558989,
+    passedProcess: '1CAL',
+    previousProcess: '1PCM',
+    nextProcess: '포장',
+    targetWidth: 772,
+    targetThickness: 0.47,
+    worktime: 48,
+  },
+  {
+    key: '3',
+    materialNumber: 'C3DL0002',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 52.9602313116795,
+    innerDiameter: 25,
+    width: 775,
+    length: 6848,
+    thickness: 0.504473434969,
+    unitWeight: 434.688234352393,
+    passedProcess: '1CAL',
+    previousProcess: '1PCM',
+    nextProcess: '포장',
+    changed: true,
+    targetWidth: 772,
+    targetThickness: 0.49,
+    worktime: 29,
+  },
+
+  {
+    key: '4',
+    materialNumber: 'C6EK0001',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 35.9824301584831,
+    innerDiameter: 22,
+    width: 775,
+    length: 2547,
+    thickness: 0.621409153683,
+    unitWeight: 68.5737101624795,
+    passedProcess: '1CAL',
+    previousProcess: '1PCM',
+    nextProcess: '포장',
+    storageLocation: '7-2-20-2-1',
+    yardDivision: '1A1B',
+    orderNumber: '20240730-F6BG',
+    targetWidth: 770,
+    targetThickness: 0.48,
+    worktime: 34,
+  },
+  {
+    key: '5',
+    materialNumber: 'C3i00003',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 49.4609835245515,
+    innerDiameter: 24,
+    width: 772,
+    length: 5876,
+    thickness: 0.620238096071,
+    unitWeight: 645.773619063763,
+    passedProcess: '1CAL',
+    previousProcess: '',
+    nextProcess: '포장',
+    targetWidth: 770,
+    targetThickness: 0.57,
+    worktime: 38,
+  },
+
+  {
+    key: '6',
+    materialNumber: 'Cbq0005',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 60.8096102448159,
+    innerDiameter: 28,
+    width: 766,
+    length: 9154,
+    thickness: 0.50512372108997,
+    unitWeight: 504.379666723385,
+    passedProcess: '1CAL',
+    previousProcess: '1PCM',
+    nextProcess: '1EGL',
+    targetWidth: 765,
+    targetThickness: 0.39,
+    worktime: 35,
+  },
+  {
+    key: '7',
+    materialNumber: 'C3Nj0004',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'A',
+    outerDiameter: 54.0121547172835,
+    innerDiameter: 28,
+    width: 769,
+    length: 6702,
+    thickness: 0.337581205233,
+    unitWeight: 666.443669362022,
+    passedProcess: '1CAL',
+    previousProcess: '1PCM',
+    nextProcess: '포장',
+    targetWidth: 765,
+    targetThickness: 0.3,
+    worktime: 30,
+  },
+];
+
+// Table Data 정의
+export const schMockdataB: MaterialDataType[] = [
+  {
+    key: '6',
+    materialNumber: 'Cow0006',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'B',
+    outerDiameter: 53.7773466191942,
+    innerDiameter: 25,
+    width: 768,
+    length: 7122,
+    thickness: 0.6426305159542,
+    unitWeight: 217.257526025889,
+    passedProcess: '1CAL',
+    previousProcess: '1PCM',
+    nextProcess: '포장',
+    targetWidth: 765,
+    targetThickness: 0.62,
+  },
+
+  {
+    key: '9',
+    materialNumber: 'CGW0009',
+    materialStatus: '2',
+    operationCode: 'C',
+    processCode: '1CAL',
+    materialType: 'C',
+    materialProgress: 'D',
+    materialGroup: 'B',
+    outerDiameter: 53.526933092068,
+    innerDiameter: 29,
+    width: 784,
+    length: 6359,
+    thickness: 0.70274970271302,
+    unitWeight: 105.774351816828,
+    passedProcess: '1PCM',
+    previousProcess: '1CAL',
+    nextProcess: '1EGL',
+    targetWidth: 765,
     targetThickness: 0.67,
   },
   {
